+++ src/app/locales/en/components/content-spaces-harvesters.js
@@ -1,0 +1,7 @@
+export default {
+  header: 'Harvesters',
+  empty: {
+    header: 'Space harvesters',
+    description: 'This space does not provide metadata to any harvester. To start indexing process, add a harvester.',
+  },
+};