+++ src/app/locales/en/components/sidebar-spaces.js
@@ -11,5 +11,6 @@ export default {
     providers: 'Providers',
     members: 'Members',
     membership: 'Membership',
+    harvesters: 'Harvesters',
   },
 };