+++ src/app/locales/en/translations.js
@@ -30,6 +30,7 @@ import contentSpacesProviders from './components/content-spaces-providers';
 import contentSpacesData from './components/content-spaces-data';
 import contentSpacesShares from './components/content-spaces-shares';
 import contentSpacesTransfers from './components/content-spaces-transfers';
+import contentSpacesHarvesters from './components/content-spaces-harvesters';
 import oneproviderViewContainer from './components/oneprovider-view-container';
 import contentProviderEmpty from './components/content-provider-empty';
 import leaveSpaceConfirm from './components/leave-space-confirm';
@@ -98,6 +99,8 @@ import guiUtils from './services/gui-utils';
 import uploadManager from './services/upload-manager';
 
 import uploadObject from './utils/upload-object';
+import addHarvesterToSpaceAction from './utils/space-actions/add-harvester-to-space-action';
+import removeHarvesterFromSpaceAction from './utils/space-actions/remove-harvester-from-space-action';
 import cleanObsoleteTokensAction from './utils/token-actions/clean-obsolete-tokens-action';
 import createTokenAction from './utils/token-actions/create-token-action';
 import openCreateTokenViewAction from './utils/token-actions/open-create-token-view-action';
@@ -149,6 +152,7 @@ const translations = {
     contentSpacesData,
     contentSpacesShares,
     contentSpacesTransfers,
+    contentSpacesHarvesters,
     oneproviderViewContainer,
     leaveSpaceConfirm,
     contentGroupsNew,
@@ -215,6 +219,10 @@ const translations = {
   },
   utils: {
     uploadObject,
+    spaceActions: {
+      addHarvesterToSpaceAction,
+      removeHarvesterFromSpaceAction,
+    },
     tokenActions: {
       cleanObsoleteTokensAction,
       createTokenAction,