+++ src/app/services/space-manager.js
@@ -9,7 +9,7 @@
 
 import Service, { inject as service } from '@ember/service';
 import { get, getProperties } from '@ember/object';
-import { Promise, resolve } from 'rsvp';
+import { resolve, all as allFulfilled } from 'rsvp';
 import ignoreForbiddenError from 'onedata-gui-common/utils/ignore-forbidden-error';
 import gri from 'onedata-gui-websocket-client/utils/gri';
 import { entityType as spaceEntityType } from 'onezone-gui/models/space';
@@ -95,7 +95,7 @@ export default Service.extend({
     return this.get('currentUser').getCurrentUserRecord()
       .then(user => user.leaveSpace(entityId))
       .then(destroyResult => {
-        return Promise.all([
+        return allFulfilled([
           this.reloadList(),
           space ? space.reload().catch(ignoreForbiddenError) : resolve(),
           this.reloadEffUserList(entityId).catch(ignoreForbiddenError),
@@ -130,7 +130,7 @@ export default Service.extend({
           subscribe: false,
         })
       )
-      .then(() => Promise.all([
+      .then(() => allFulfilled([
         space ? space.reload() : resolve(),
         this.reloadUserList(entityId).catch(ignoreForbiddenError),
         this.reloadEffUserList(entityId).catch(ignoreForbiddenError),
@@ -157,7 +157,7 @@ export default Service.extend({
         data: childGroupRepresentation,
         authHint: ['asUser', get(user, 'entityId')],
       }).then(() => {
-        return Promise.all([
+        return allFulfilled([
           this.reloadGroupList(spaceEntityId).catch(ignoreForbiddenError),
           this.reloadEffGroupList(spaceEntityId).catch(ignoreForbiddenError),
           this.get('groupManager').reloadList(),
@@ -182,7 +182,7 @@ export default Service.extend({
       }),
       operation: 'create',
     }).then(() => {
-      return Promise.all([
+      return allFulfilled([
         this.reloadGroupList(spaceEntityId).catch(ignoreForbiddenError),
         this.reloadEffGroupList(spaceEntityId).catch(ignoreForbiddenError),
         this.get('groupManager').reloadSpaceList(groupEntityId)
@@ -205,10 +205,10 @@ export default Service.extend({
       'user',
       userEntityId
     ).then(() =>
-      Promise.all([
+      allFulfilled([
         this.reloadUserList(spaceEntityId).catch(ignoreForbiddenError),
         this.reloadEffUserList(spaceEntityId).catch(ignoreForbiddenError),
-        currentUser.runIfThisUser(userEntityId, () => Promise.all([
+        currentUser.runIfThisUser(userEntityId, () => allFulfilled([
           this.reloadList(),
           this.get('providerManager').reloadList(),
           space ? space.reload().catch(ignoreForbiddenError) : resolve(),
@@ -229,7 +229,7 @@ export default Service.extend({
       spaceEntityType,
       spaceEntityId
     ).then(() =>
-      Promise.all([
+      allFulfilled([
         this.reloadGroupList(spaceEntityId).catch(ignoreForbiddenError),
         this.reloadList(),
         this.get('providerManager').reloadList(),
@@ -251,7 +251,7 @@ export default Service.extend({
       'group',
       groupEntityId
     ).then(() =>
-      Promise.all([
+      allFulfilled([
         this.reloadGroupList(spaceEntityId).catch(ignoreForbiddenError),
         this.reloadEffGroupList(spaceEntityId).catch(ignoreForbiddenError),
         this.reloadList(),
@@ -260,21 +260,6 @@ export default Service.extend({
         .catch(ignoreForbiddenError),
       ])
     );
-  },
-
-  /**
-   * Joins space to a harvester using token
-   * @param {Model.Space} space 
-   * @param {string} token
-   * @returns {Promise<Model.Harvester>}
-   */
-  joinSpaceToHarvester(space, token) {
-    const harvesterManager = this.get('harvesterManager');
-    return space.joinHarvester(token)
-      .then(harvester => harvesterManager.reloadSpaceList(get(harvester, 'entityId'))
-        .catch(ignoreForbiddenError)
-        .then(() => space)
-      );
   },
 
   /**